requirements.txt
utils~=1.0.1
future~=0.18.2
Pillow~=9.0.1
google~=3.0.0
setuptools~=41.6.0
keyboard~=0.13.5
pyttsx3~=2.7
beautifulsoup4~=4.12.3